Feature (rag): Add DOC file loader
-Added ApachePoiDocLoader class for loading DOC files
-Implemented text and image extraction functions for DOC files - added unit testing for DOC file loader
-Refactored the original ApachePoiLoader class and renamed it ApachePoiDocxLoader
-Updated the relevant test class names
