+++ pom.xml
@@ -207,6 +207,12 @@
       <version>5.2.5</version>
     </dependency>
 
+    <dependency>
+      <groupId>org.apache.poi</groupId>
+      <artifactId>poi-scratchpad</artifactId>
+      <version>5.2.3</version>
+    </dependency>
+
   </dependencies>
 
   <build>
