+++ src/main/java/org/salt/jlangchain/rag/loader/doc/BaseDocLoader.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.salt.jlangchain.rag.loader.doc;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.experimental.SuperBuilder;
+import org.salt.jlangchain.rag.loader.BaseLoader;
+import org.salt.jlangchain.rag.loader.ocr.OcrActuator;
+import org.salt.jlangchain.rag.loader.ocr.TesseractActuator;
+
+import java.io.InputStream;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@SuperBuilder
+public abstract class BaseDocLoader extends BaseLoader {
+
+    protected String filePath;
+    protected String webPath;
+    protected InputStream inputStream;
+    @Builder.Default
+    protected boolean extractImages = false;
+    @Builder.Default
+    protected Class<? extends OcrActuator> ocrClazz = TesseractActuator.class;
+}
